+++ src/libraries/ui.js
@@ -36,6 +36,7 @@ import taskInventoryHighlighter from '../features/tasks/task-inventory-highlight
 import taskStatistics from '../features/tasks/task-statistics.js';
 import taskClaimCollector from '../features/tasks/task-claim-collector.js';
 import taskRerollProtection from '../features/tasks/task-reroll-protection.js';
+import taskBulkReroll from '../features/tasks/task-bulk-reroll.js';
 import taskAutoReroll from '../features/tasks/task-auto-reroll.js';
 
 // Skills
@@ -119,6 +120,7 @@ toolashaRoot.UI = {
     taskStatistics,
     taskClaimCollector,
     taskRerollProtection,
+    taskBulkReroll,
     taskAutoReroll,
     remainingXP,
     xpTracker,